stan: constrain Gamma shape and (inverse) scale > 0
Updated discretised_gamma_pmf (discretised truncated Gamma PMF) to constrain gamma shape and (inverse) scale parameters to be positive and finite (alpha > 0 and beta > 0).

Fixing the following message:

Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
Chain 1 Exception: Exception: gamma_cdf: Shape parameter is 0, but must be positive and finite!