Ensure unbiased summary of growth rate based on infectiousness towards present
This fix accounts for the fact that the growth rate on the most recent dates is undefined by the mean of the generation time, but since the mean generation time is estimated, it varies across draws. The standard summary used by epinow can then lead to biased growth rate estimates for dates where some draws still have a growth rate and other not.