Growth estimators (#1076)
* Add new calculate_growth method

* Add option growth_method to estimate_infections

* Add growth_method to simulate_infections

* Fix growth_method call in simulate_infections

Co-authored-by: Sebastian Funk <sebastian.funk@lshtm.ac.uk>

* set most recent growth rates to not_a_number()

* Add empty newline at end of helper.stan

* Add doc for growth_method in rt_opts

* Add wrapper for calculate_growth

* Update rt_opts and simulate_infections doc

* update further docs (hitchhiking)

* Add helpers to expose_stan_fns

* Remove stats::convolve (not used anymore)

* lint

* Add tests for calculate growth methods

* Add NEWS

* Add author

* Minor fixes NEWS and package doc

* Refactor calculate_growth_infections

Now, the shift due to differencing is internally accounted for (not via seeding_time + 1 anymore).

* Refactor vector calculate_growth_infness

More verbatim code for shifting, and shift from differencing internally accounted for

* Ensure unbiased summary of growth rate based on infectiousness towards present

This fix accounts for the fact that the growth rate on the most recent dates is undefined by the mean of the generation time, but since the mean generation time is estimated, it varies across draws. The standard summary used by epinow can then lead to biased growth rate estimates for dates where some draws still have a growth rate and other not.

* Add integration test growth method infectiousness

* More linting

---------

Co-authored-by: Sebastian Funk <sebastian.funk@lshtm.ac.uk>